+++ graphics/src/geom.rs
@@ -3,39 +3,39 @@ use std::ops;
 /// A 2-dimensional vector.
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Vec2 {
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Vec2 {
     /// Constructs a 2-dimensional vector.
     #[inline]
-    pub fn new(x: f64, y: f64) -> Vec2 {
+    pub fn new(x: f32, y: f32) -> Vec2 {
         Vec2 { x: x, y: y }
     }
 
     /// Computes the dot product between two vectors.
     #[inline]
-    pub fn dot(self, other: Vec2) -> f64 {
+    pub fn dot(self, other: Vec2) -> f32 {
         self.x * other.x + self.y * other.y
     }
 
     /// Considering the two given vectors as 3-dimensional vectors lying in the
     /// XY-plane, finds the z-coordinate of their cross product.
     #[inline]
-    pub fn cross(self, other: Vec2) -> f64 {
+    pub fn cross(self, other: Vec2) -> f32 {
         self.x * other.y - self.y * other.x
     }
 
     /// Computes the distance between two points.
     #[inline]
-    pub fn distance(self, other: Vec2) -> f64 {
+    pub fn distance(self, other: Vec2) -> f32 {
         (other - self).length()
     }
 
     /// Computes the length of a vector.
     #[inline]
-    pub fn length(self) -> f64 {
+    pub fn length(self) -> f32 {
         self.dot(self).sqrt()
     }
 
@@ -47,7 +47,7 @@ impl Vec2 {
 
     /// Linearly interpolates between two vectors by the parameter `t`.
     #[inline]
-    pub fn lerp(t: f64, a: Vec2, b: Vec2) -> Vec2 {
+    pub fn lerp(t: f32, a: Vec2, b: Vec2) -> Vec2 {
         (1.0 - t) * a + t * b
     }
 
@@ -94,15 +94,15 @@ impl ops::SubAssign for Vec2 {
     }
 }
 
-impl ops::Mul<f64> for Vec2 {
+impl ops::Mul<f32> for Vec2 {
     type Output = Vec2;
     #[inline]
-    fn mul(self, rhs: f64) -> Vec2 {
+    fn mul(self, rhs: f32) -> Vec2 {
         Vec2 { x: self.x * rhs, y: self.y * rhs }
     }
 }
 
-impl ops::Mul<Vec2> for f64 {
+impl ops::Mul<Vec2> for f32 {
     type Output = Vec2;
     #[inline]
     fn mul(self, rhs: Vec2) -> Vec2 {
@@ -112,11 +112,11 @@ impl ops::Mul<Vec2> for f64 {
 
 /// A 2×2 matrix, in row-major order.
 #[derive(Copy, Clone, Debug, PartialEq)]
-pub struct Mat2x2(pub [f64; 4]);
+pub struct Mat2x2(pub [f32; 4]);
 
 impl Mat2x2 {
     /// Constructs a 2×2 matrix. Arguments are given in row-major order.
-    pub fn new(a: f64, b: f64, c: f64, d: f64) -> Mat2x2 {
+    pub fn new(a: f32, b: f32, c: f32, d: f32) -> Mat2x2 {
         Mat2x2([a, b, c, d])
     }
 
@@ -126,12 +126,12 @@ impl Mat2x2 {
     }
 
     /// Constructs a uniform scaling matrix.
-    pub fn scale(scale: f64) -> Mat2x2 {
+    pub fn scale(scale: f32) -> Mat2x2 {
         Mat2x2([scale, 0.0, 0.0, scale])
     }
 
     /// Constructs a rotation matrix.
-    pub fn rotate(angle: f64) -> Mat2x2 {
+    pub fn rotate(angle: f32) -> Mat2x2 {
         Mat2x2([angle.cos(), angle.sin(), -angle.sin(), angle.cos()])
     }
 }
@@ -157,7 +157,7 @@ impl ops::Mul<Vec2> for Mat2x2 {
     }
 }
 
-impl ops::Mul<Mat2x2> for f64 {
+impl ops::Mul<Mat2x2> for f32 {
     type Output = Mat2x2;
     #[inline]
     fn mul(self, rhs: Mat2x2) -> Mat2x2 {
@@ -165,10 +165,10 @@ impl ops::Mul<Mat2x2> for f64 {
     }
 }
 
-impl ops::Mul<f64> for Mat2x2 {
+impl ops::Mul<f32> for Mat2x2 {
     type Output = Mat2x2;
     #[inline]
-    fn mul(self, rhs: f64) -> Mat2x2 {
+    fn mul(self, rhs: f32) -> Mat2x2 {
         rhs * self
     }
 }
@@ -193,17 +193,17 @@ impl Transform {
     }
 
     /// Constructs a translation.
-    pub fn translate(x: f64, y: f64) -> Transform {
+    pub fn translate(x: f32, y: f32) -> Transform {
         Transform { matrix: Mat2x2::id(), offset: Vec2::new(x, y) }
     }
 
     /// Constructs a uniform scaling.
-    pub fn scale(scale: f64) -> Transform {
+    pub fn scale(scale: f32) -> Transform {
         Transform { matrix: Mat2x2::scale(scale), offset: Vec2::new(0.0, 0.0) }
     }
 
     /// Constructs a rotation.
-    pub fn rotate(angle: f64) -> Transform {
+    pub fn rotate(angle: f32) -> Transform {
         Transform { matrix: Mat2x2::rotate(angle), offset: Vec2::new(0.0, 0.0) }
     }
